Move AllocatorConfig to impl
