--- tt_metal/impl/allocator/allcator_types.hpp
@@ -1,18 +1,0 @@
-// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#pragma once
-
-namespace tt::tt_metal {
-
-// Setup what each core-type is
-enum class AllocCoreType {
-    Dispatch,
-    StorageOnly,
-    ComputeOnly,
-    ComputeAndStore,
-    Invalid,
-};
-
-};  // namespace tt::tt_metal
