+++ tt_metal/impl/allocator/l1_banking_allocator.cpp
@@ -24,7 +24,7 @@
 #include "bank_manager.hpp"
 #include "hal_types.hpp"
 #include "impl/context/metal_context.hpp"
-#include "impl/allocator/allcator_types.hpp"
+#include "impl/allocator/allocator_types.hpp"
 #include <umd/device/types/xy_pair.hpp>
 #include <tt-metalium/tt_align.hpp>
